+++ src/assets/release_notes.md
@@ -1,4 +1,13 @@
 ## Release Notes
+## v2.0.6 - 23/03/2021
+### Improvements <img class="joypixels" width="18" height="18" src="https://cdn.jsdelivr.net/joypixels/assets/6.5/png/unicode/32/1f528.png">
+- Add album: Automatically set the oldest date when adding pictures
+- Collection selector: A more convenient collection display is implemented 
+
+### Bug Fixes <img class="joypixels" width="18" height="18" src="https://cdn.jsdelivr.net/joypixels/assets/6.5/png/unicode/32/1f41b.png">
+- Legacy import: Fixed the dialog dimensions
+- Start organizing: Fixed renaming albums that exist of multiple words
+
 ## v2.0.5 - 09/03/2021
 ### Improvements <img class="joypixels" width="18" height="18" src="https://cdn.jsdelivr.net/joypixels/assets/6.5/png/unicode/32/1f528.png">
 - Releases notes are displayed to the user when a new version of the application is detected